Proposal: Extend from ActiveModel
I was running into bugs in the ember-data-endpoints adapter that I
wasn't running into using ActiveModelAdapter (such as date fields
not normalizing, etc.).

I propose we extend from ActiveModel, which, except for pathForType,
and serializing, works with our API.

The empty extend serializer in this commit will be necessary for
adding the serializeIntoHash method to make it nest-api-compat.